Move chart previews into rule types folder
